requirements.txt
deepspeed==0.15.4
peft==0.11.1
bitsandbytes==0.43.1
scipy
datasets==2.20.0
safetensors==0.4.3
scikit-learn
prettytable
pillow
pandas
fire
wandb
pydantic==2.10.6
qwen_vl_utils
transformers
accelerate
